some refactoring user destroy, store and test

+++ back-end/app/controllers/users_controller.ts
@@ -9,6 +9,7 @@ import {
 } from '#validators/user'
 import db from '@adonisjs/lucid/services/db'
 import { DateTime } from 'luxon'
+import { Exception } from '@adonisjs/core/exceptions'
 
 export default class UsersController {
   async index({ request }: HttpContext) {
@@ -18,33 +19,26 @@ export default class UsersController {
   }
 
   async store({ request, auth, response }: HttpContext) {
-    // check if email already exists in our database.
     const { email } = request.only(['email'])
-
     const user = await User.findBy('email', email)
-    // isPersisted true, user exist
-
     if (user?.$isPersisted) {
-      return response.status(400).json({ message: 'user already exists in our database' })
+      return response.badRequest({ message: 'user already exists in our database' })
     }
 
-    //next auth
     if (auth.user?.isAdmin) {
       const payload = await request.validateUsing(InternalUserValidator)
-      return await User.create(payload)
+      const createdUser = await User.create(payload)
+      return response.created({ message: 'user creation sucessed', user: createdUser })
     } else {
-      // response.status(201).json({ message: 'user creation sucessed' })
-
-      const newUserData = Object.assign(request.all())
-      const payload = await ExternalUserValidator.validate(newUserData) //status 422 error, bad request
+      const payload = await request.validateUsing(ExternalUserValidator)
       const user = await User.create(Object.assign(payload, { isAdmin: false }))
       const token = await User.accessTokens.create(user)
-      return { token: token, user: user }
+      return response.created({ message: 'user creation sucessed', token: token, user: user })
     }
   }
 
   async update({ params, request, auth }: HttpContext) {
-    const user = await User.findOrFail(params.id) //error 500
+    const user = await User.findOrFail(params.id)
     // update update at
     const updatedAt = { updatedAt: DateTime.local() }
     const { isProfile } = request.only(['isProfile'])
@@ -90,19 +84,15 @@ export default class UsersController {
     }
   }
 
-  async destroy({ params, auth }: HttpContext) {
-    //gets both parameter id and user id.
-    const parameter = params.id
-    const userId = auth.user?.id! as number
-    // admin deleting users
-    if (auth.user?.isAdmin) {
-      // return { params: params, typeofparams: typeof params.id }
-      const user = await User.findOrFail(parameter) //status 404 error, not found
-      await user.delete()
-    } //guest deleting itself.
-    else if (!auth.user?.isAdmin && parseInt(parameter) === auth.user?.id) {
-      await db.from('auth_access_tokens').where('tokenable_id', userId).delete()
-      await db.from('users').where('id', userId).delete()
+  async destroy({ params, auth, response }: HttpContext) {
+    if (!auth.user?.isAdmin && auth.user?.id != params.id) {
+      return response.unauthorized(
+        'guest should not be able to delete another user, beside itself.'
+      )
     }
+    const user = await User.findOrFail(params.id)
+    await db.from('auth_access_tokens').where('tokenable_id', auth.user?.id!).delete()
+    await user.delete()
+    return response.noContent()
   }
 }
